#184 Implement discrete rounding (no-op)
